requirements.txt
attrs==23.2.0
certifi==2024.2.2
h11==0.14.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.0
pluggy==1.4.0
PySocks==1.7.1
pytest==8.1.1
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.0.1
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
